+++ tests/e2e_tests/test_profiler.py
@@ -58,6 +58,7 @@ def test_full_run():
         "name: Eric Matthes",
         "mastodon: https://fosstodon.org/@ehmatthes",
         "Empty fields: company, bio",
+        "Orgs: openlearningtools, django-simple-deploy",
     )
 
     for expected_string in expected_strings: