+++ tests/unit_tests/reference_files/reference_summaries.py
@@ -14,13 +14,18 @@ GitHub user: ehmatthes
 🟢 No concerns found with recent PR activity.
    5 PRs opened in the last 21 days.
       0 opened against repos the user owns.
+      0 opened against repos in publicly associated orgs.
       5 opened against external repos.
 
    🟢 3 of 5 external PRs merged in the last 21 days.
    🟢 1 of 5 external PRs closed without merging in the last 21 days.
 
 🟢 No concerns found with recent issue activity.
-   🟢 7 new issues opened in the last 21 days.
-   🟢 0 issues closed as NOT_PLANNED.
-   🟢 0 issues opened with the same title.
+   7 new issues opened in the last 21 days.
+      7 opened in repos the user owns.
+      0 opened in repos in publicly associated orgs.
+      0 opened in external repos.
+
+   🟢 0 external issues closed as NOT_PLANNED.
+   🟢 0 external issues opened with the same title.
 """.strip()